Cleanup - get rid of the NSInteger, NSNumber, stuff.
Just use int until you need it.

+++ OCDSpec/OCDSpecFail.h
@@ -5,7 +5,7 @@
 
 }
 
-+(void) fail:(NSString*)reason atLine:(NSInteger) line inFile: (NSString *)file;
++(void) fail:(NSString*)reason atLine:(int) line inFile: (NSString *)file;
 
 @end
