+++ OCDSpec/OCDSpecFail.m
@@ -2,7 +2,7 @@
 
 @implementation OCDSpecFail
 
-+(void) fail:(NSString *)reason atLine:(NSInteger) line inFile: (NSString *)file
++(void) fail:(NSString *)reason atLine:(int) line inFile: (NSString *)file
 {
   NSException *exception = [NSException exceptionWithName:@"Test Failed"
                                                    reason:reason
